+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/sweep/metrics/TargetedSweepMetricsTest.java
@@ -27,6 +27,10 @@ import com.codahale.metrics.Gauge;
 import com.codahale.metrics.MetricRegistry;
 import com.google.common.collect.ImmutableMap;
 import com.palantir.atlasdb.AtlasDbMetricNames;
+import com.palantir.atlasdb.cleaner.KeyValueServicePuncherStore;
+import com.palantir.atlasdb.cleaner.PuncherStore;
+import com.palantir.atlasdb.keyvalue.api.KeyValueService;
+import com.palantir.atlasdb.keyvalue.impl.InMemoryKeyValueService;
 import com.palantir.atlasdb.sweep.queue.ShardAndStrategy;
 import com.palantir.atlasdb.util.AtlasDbMetrics;
 import com.palantir.tritium.metrics.registry.MetricName;
@@ -36,11 +40,16 @@ public class TargetedSweepMetricsTest {
     private static final ShardAndStrategy CONS_ONE = ShardAndStrategy.conservative(1);
     private static final ShardAndStrategy CONS_TWO = ShardAndStrategy.conservative(2);
     private static final ShardAndStrategy THOR_ZERO = ShardAndStrategy.thorough(0);
+    private PuncherStore puncherStore;
+    private long clockTime;
     private TargetedSweepMetrics metrics;
 
     @Before
     public void setup() {
-        metrics = TargetedSweepMetrics.withRecomputingInterval(1);
+        clockTime = 100;
+        KeyValueService kvs = new InMemoryKeyValueService(true);
+        puncherStore = KeyValueServicePuncherStore.create(kvs, false);
+        metrics = TargetedSweepMetrics.createWithClock(kvs, () -> clockTime, 1);
     }
 
     @Test
@@ -58,7 +67,12 @@ public class TargetedSweepMetricsTest {
         assertTombstonesPutConservativeEquals(1);
         assertAbortedWritesDeletedConservativeEquals(2);
         assertSweepTimestampConservativeEquals(7);
-        assertLastSweptConservativeEquals(4);
+        assertLastSweptTimestampConservativeEquals(4);
+
+        puncherStore.put(3, 2);
+        puncherStore.put(4, 4);
+        puncherStore.put(5, 40);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 4);
     }
 
     @Test
@@ -76,7 +90,12 @@ public class TargetedSweepMetricsTest {
         assertTombstonesPutThoroughEquals(2);
         assertAbortedWritesDeletedThoroughEquals(3);
         assertSweepTimestampThoroughEquals(9);
-        assertLastSweptThoroughEquals(6);
+        assertLastSweptTimestampThoroughEquals(6);
+
+        puncherStore.put(5, 1);
+        puncherStore.put(6, 6);
+        puncherStore.put(7, 16);
+        assertMillisSinceLastSweptThoroughForClockTimeEquals(clockTime, 6);
     }
 
     @Test
@@ -134,18 +153,55 @@ public class TargetedSweepMetricsTest {
         metrics.updateProgressForShard(CONS_TWO, 1000);
         waitForProgressToRecompute();
 
-        assertLastSweptConservativeEquals(1);
+        assertLastSweptTimestampConservativeEquals(1);
+
+        puncherStore.put(0, 5);
+        puncherStore.put(2, 500);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 5);
+    }
+
+    @Test
+    public void millisSinceLastSweptUpdatesAsClockUpdates() {
+        metrics.updateProgressForShard(CONS_ZERO, 100);
+        waitForProgressToRecompute();
+
+        puncherStore.put(0, 50);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 50);
+
+        for (int i = 0; i < 10; i++) {
+            clockTime = i;
+            assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 50);
+        }
+    }
+
+    @Test
+    public void millisSinceLastSweptDoesNotReadPuncherAgainUntilLastSweptDoesNotChange() {
+        metrics.updateProgressForShard(CONS_ZERO, 100);
+        waitForProgressToRecompute();
+
+        puncherStore.put(0, 50);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 50);
+        puncherStore.put(1, 100);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 50);
+
+        metrics.updateProgressForShard(CONS_ZERO, 101);
+        waitForProgressToRecompute();
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 100);
     }
 
     @Test
     public void lastSweptGoesDownIfNewInformationBecomesAvailable() {
         metrics.updateProgressForShard(CONS_ZERO, 999);
         waitForProgressToRecompute();
-        assertLastSweptConservativeEquals(999);
+        assertLastSweptTimestampConservativeEquals(999);
+        puncherStore.put(999, 999);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 999);
 
         metrics.updateProgressForShard(CONS_ONE, 200);
         waitForProgressToRecompute();
-        assertLastSweptConservativeEquals(200);
+        assertLastSweptTimestampConservativeEquals(200);
+        puncherStore.put(200, 200);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 200);
     }
 
     @Test
@@ -158,7 +214,10 @@ public class TargetedSweepMetricsTest {
         metrics.updateProgressForShard(CONS_ONE, 150);
         waitForProgressToRecompute();
 
-        assertLastSweptConservativeEquals(100);
+        assertLastSweptTimestampConservativeEquals(100);
+        puncherStore.put(1, 1);
+        puncherStore.put(100, 100);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 100);
     }
 
     @Test
@@ -172,12 +231,16 @@ public class TargetedSweepMetricsTest {
         metrics.updateProgressForShard(CONS_TWO, 50);
         waitForProgressToRecompute();
 
-        assertLastSweptConservativeEquals(10);
+        assertLastSweptTimestampConservativeEquals(10);
+        puncherStore.put(10, 10);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 10);
 
         metrics.updateProgressForShard(CONS_ONE, 40);
         waitForProgressToRecompute();
 
-        assertLastSweptConservativeEquals(30);
+        assertLastSweptTimestampConservativeEquals(30);
+        puncherStore.put(30, 30);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 30);
     }
 
 
@@ -231,16 +294,23 @@ public class TargetedSweepMetricsTest {
         metrics.updateProgressForShard(THOR_ZERO, 50);
         waitForProgressToRecompute();
 
-        assertLastSweptConservativeEquals(1);
-        assertLastSweptThoroughEquals(50);
+        assertLastSweptTimestampConservativeEquals(1);
+        assertLastSweptTimestampThoroughEquals(50);
+        puncherStore.put(1, 1);
+        puncherStore.put(50, 50);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 1);
+        assertMillisSinceLastSweptThoroughForClockTimeEquals(clockTime, 50);
 
         metrics.updateProgressForShard(CONS_ZERO, 10);
         metrics.updateProgressForShard(CONS_ONE, 5);
         metrics.updateProgressForShard(THOR_ZERO, 5);
         waitForProgressToRecompute();
 
-        assertLastSweptConservativeEquals(5);
-        assertLastSweptThoroughEquals(5);
+        puncherStore.put(5, 5);
+        assertLastSweptTimestampConservativeEquals(5);
+        assertLastSweptTimestampThoroughEquals(5);
+        assertMillisSinceLastSweptConservativeForClockTimeEquals(clockTime, 5);
+        assertMillisSinceLastSweptThoroughForClockTimeEquals(clockTime, 5);
     }
 
     public static void waitForProgressToRecompute() {
@@ -271,8 +341,16 @@ public class TargetedSweepMetricsTest {
         assertThat(getGaugeConservative(AtlasDbMetricNames.SWEEP_TS).getValue()).isEqualTo(value);
     }
 
-    public static void assertLastSweptConservativeEquals(long value) {
-        assertThat(getGaugeConservative(AtlasDbMetricNames.SWEPT_TS).getValue()).isEqualTo(value);
+    public static void assertLastSweptTimestampConservativeEquals(long value) {
+        assertThat(getGaugeConservative(AtlasDbMetricNames.LAST_SWEPT_TS).getValue()).isEqualTo(value);
+    }
+
+    private static void assertMillisSinceLastSweptConservativeForClockTimeEquals(long currentTime, long lastSwept) {
+        assertThat(getGaugeConservative(AtlasDbMetricNames.LAG_MILLIS).getValue()).isEqualTo(currentTime - lastSwept);
+    }
+
+    public static void assertMillisSinceLastSweptConservativeLessThanOneSecond() {
+        assertThat(getGaugeConservative(AtlasDbMetricNames.LAG_MILLIS).getValue()).isBetween(0L, 1000L);
     }
 
     public static Gauge<Long> getGaugeConservative(String name) {
@@ -299,8 +377,12 @@ public class TargetedSweepMetricsTest {
         assertThat(getGaugeThorough(AtlasDbMetricNames.SWEEP_TS).getValue()).isEqualTo(value);
     }
 
-    public static void assertLastSweptThoroughEquals(long value) {
-        assertThat(getGaugeThorough(AtlasDbMetricNames.SWEPT_TS).getValue()).isEqualTo(value);
+    public static void assertLastSweptTimestampThoroughEquals(long value) {
+        assertThat(getGaugeThorough(AtlasDbMetricNames.LAST_SWEPT_TS).getValue()).isEqualTo(value);
+    }
+
+    public static void assertMillisSinceLastSweptThoroughForClockTimeEquals(long currentTime, long lastSwept) {
+        assertThat(getGaugeThorough(AtlasDbMetricNames.LAG_MILLIS).getValue()).isEqualTo(currentTime - lastSwept);
     }
 
     public static Gauge<Long> getGaugeThorough(String name) {